code/mcp/requirements.txt
fastapi==0.128.4
uvicorn[standard]==0.40.0
mcp==1.26.0
python-dotenv==1.2.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-mcp==0.52.1
